refactor: reorganize chatting service and speech-pipeline modules
chatting-service/ → chatting/
audio-pipeline/ → speech-pipeline/
audio-pipeline.ts → orchestrator.ts
