+++ src/routes/api/ws/chat/index.ts
@@ -1,5 +1,5 @@
 import fastifyWebsocket from '@fastify/websocket';
-import { ChattingService } from '#@/services/chatting-service/chatting-service';
+import { ChattingService } from '#@/services/chatting/chatting.service';
 import { WS_STATUS } from '#@/constants/ws-status';
 
 import type { FastifyPluginAsync, FastifyRequest } from 'fastify';
@@ -11,7 +11,7 @@ const plugin: FastifyPluginAsync = async (fastify) => {
   fastify.get('/', { websocket: true }, async (socket: WebSocket, req: FastifyRequest) => {
     try {
       const chattingService = new ChattingService(socket, req.log);
-      chattingService.startSession();
+      await chattingService.startSession();
     } catch (err) {
       fastify.log.error(err);
       socket.close(WS_STATUS.INTERNAL.CODE, WS_STATUS.INTERNAL.MESSAGE);
